perf(api): load react-email lazily — full-graph eval 400→260ms, startup heap 37→26MB
The 2.76MB 'registry' commons chunk was ~2.1MB of @react-email/tailwind,
@react-email/code-block (prism), tailwindcss and prettier, statically
reachable from AlertDeliveryDispatch → alert-email and DigestService via
service-graph. Render sites now dynamic-import @react-email/components and
the email components; deriveDigestStatus + digest prop types move to a
react-free @maple/email/weekly-digest-core so the subject-line path stays
static.

Scorecard: eval 241/400 → 116/260 ms, heap 37.3/196.3 → 25.7/180.1 MB,
registry chunk 2.76 → 0.99 MB.
